==> 1-cah_extract_data.ipynb <==
### Extract data & train/test sets 
* pretrain data: all black + which + combinations
    * Could supplemenet with more black/white from internet, and other jokes sources (reddit, jester) but overkill. 
    * Main goal is just to "warm up" a language/sentence model, generically. 

* train/test sets: Split by game/`fake_round_id`
    * Train data: Drop duplicates
    * keep "all" 1 cases, and keep all # occurrences by default

* IGNORE pick 2 cases for now (overcomplicates evaluation) ; but keep them for the pretraining

Each row of the file represents a white card that was given to a player in a round and how it performed. It also contains information about the round like how long it took and what black card was dealt.
So columns that describe general round data will be exactly the same for all ten rows of that round. These rows are: fake_round_id, round_completion_seconds, round_skipped, black_card_text, and black_card_pick_num.
But columns that describe a property of the white card will change over the ten rows of that round. These rows are: white_card_text, won, and winning_index.


The columns:
* fake_round_id: a unique integer for every round. This is to help you separate one round from another to find which 10 cards were presented in every round.
* round_completion_seconds: the number of seconds it took for the user to pick a winner after being presented the cards for that round. You'll probably want to filter out users that decide too quickly by setting some minimum threshold for this. Some of these numbers can be insanely high as well since the user just left their browser open for hours or possibly days before picking a winner.
* round_skipped: we have a red button in the top right of the lab page that says "No Good Plays!". If a user clicks this button, round_skipped will be "True" and no white card will be marked as a winner.
* black_card_text: self explanatory.
* black_card_pick_num: The number of white cards that the black card requires. Generally 1, but will be 2 for black cards like "That's right, I killed _____. How, you ask? _____."
* white_card_text: self explanatory.
* won: True if this white card won the round.
* winning_index: This is for storing the order that the winning card(s) were picked in. If the white card didn't win, this will just contain a null value (NULL). If the card won and the black card was a normal "Pick 1" this will always be 0. But if the black card was a "Pick 2", one of the winners will be 0 and the other will be 1. This is how you figure out which blank in pick 2 black cards was meant to be filled by which winning white card. The first blank is filled by the 0 and the second one by the 1.

* sample groups (for test);

* https://stackoverflow.com/questions/44007496/random-sampling-with-pandas-data-frame-disjoint-groups
    ```
    from sklearn.model_selection import GroupShuffleSplit

    # Initialize the GroupShuffleSplit.
    gss = GroupShuffleSplit(n_splits=1, test_size=0.5)

    # Get the indexers for the split.
    idx1, idx2 = next(gss.split(df, groups=df.ids))

    # Get the split DataFrames.
    df1, df2 = df.iloc[idx1], df.iloc[idx2]
    ```

* https://hippocampus-garden.com/pandas_group_sample/
    ```
    sampled_users = np.random.choice(df["user_id"].unique(), 100)
    df_sampled = df.groupby('user_id').filter(lambda x: x["user_id"].values[0] in sampled_users)
    df_sampled["user_id"].nunique()
    ```

import pandas as pd
import seaborn as sns
pd.set_option("display.precision", 4)
pd.set_option('display.width', 1024)
pd.set_option('display.max_colwidth', None)
%matplotlib inline 
from sklearn.linear_model import LogisticRegression, LogisticRegressionCV
from sklearn.model_selection import train_test_split,GridSearchCV, cross_val_score, cross_val_predict
from sklearn.preprocessing import OrdinalEncoder, LabelEncoder
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import make_pipeline
from sklearn.metrics import classification_report, roc_auc_score  
# https://stackoverflow.com/questions/53784971/how-to-disable-convergencewarning-using-sklearn
from warnings import simplefilter
from sklearn.exceptions import ConvergenceWarning
simplefilter("ignore", category=ConvergenceWarning)
from sklearn.model_selection import GroupShuffleSplit

ord_enc = OrdinalEncoder(handle_unknown="use_encoded_value",unknown_value=-1)

# from sentence_transformers import SentenceTransformer, util ## versions error? 

LOAD_CSV = False#True#False ## load raw data from csv instead of parquet (which has had some cols dropped)
MIN_PAIR_FREQ = 1
SAVE_TRAIN_DATA = False#False#True

LOWERCASE_DATA = False
DROP_DOUBLES = True ## drop white cards with pick 2 options ; and their blacks. (TODO: analyze interesting pairs)
DROP_SKIPPED = True ## ignore rounds where skipped

DATA_PATH = #REDACTED

Load data

* Note there are some near-duplicate cards (e.g. footballers - could minhash - https://github.com/chrisjmccormick/MinHash/blob/master/runMinHashExample.py , LSH/shingles : https://onestopdataanalysis.com/lsh/  (broken library?) , https://github.com/zyocum/dedup/blob/master/dedup.py ,

`Datasketch` https://www.learndatasci.com/tutorials/building-recommendation-engine-locality-sensitive-hashing-lsh-python/ ,
https://stackoverflow.com/questions/25114338/approximate-string-matching-using-lsh , 

SKLearn friendly?: http://ethen8181.github.io/machine-learning/recsys/content_based/lsh_text.html

`dedupe` library - https://github.com/dedupeio/dedupe

* Paraphrase mining using sentence bert - https://www.sbert.net/examples/applications/paraphrase-mining/README.html  (may be too semantic)

* Example: "10 football players with erections *barreling* towards you at full speed" - 10 football players with erections **barrelling** towards you at full speed."

Task modelling:
* Ranking = most accurate
* Regression/classification: can be done just over single cards, or between pairs. 
    * if pairs-wise: can use sentencebert/SNLI models (which assume pairs of sentences) - https://github.com/UKPLab/sentence-transformers
    
    
Examples:

* https://www.sbert.net/examples/training/sts/README.html#training-data
* https://github.com/UKPLab/sentence-transformers/blob/master/examples/training/nli/training_nli.py
* https://github.com/UKPLab/sentence-transformers/blob/master/examples/training/nli/training_nli_v2.py


* Keras example (not sentenceBert): https://keras.io/examples/nlp/semantic_similarity_with_bert/


* Interpretability tool : https://github.com/cdpierse/transformers-interpret

%%time
#### Original processing of data, before I saved it as a parquet for faster subsequent loading: 
if LOAD_CSV:
    df = pd.read_csv(DATA_PATH)
    ## remove punct/newlines
    df["white_card_text"] = df["white_card_text"].replace("\n", " \n ",regex=True).replace("\t", " \t ",regex=True) # .replace("\r", "")
    df["black_card_text"] = df["black_card_text"].replace("\n", " \n ",regex=True).replace("\t", " \t ",regex=True) # .replace("\r", "")

    ## add number within group - may not actually be the read order shown on screen!! 
    df['ID_index'] = df.groupby('fake_round_id').cumcount() + 1

    print(df[['black_card_text', 'black_card_pick_num', 'white_card_text',"fake_round_id"]].nunique())
    print(df.describe())
    
    ## save to disk in compressed, binary format
    df.drop(["winning_index"],axis=1).to_parquet('cah_lab_data.parquet')

else:  # load data saved in above code previously
    df = pd.read_parquet('cah_lab_data.parquet')
#     df= pd.read_feather('cah_lab_data.feather')

if DROP_DOUBLES:
    df = df.loc[df["black_card_pick_num"]==1]
if DROP_SKIPPED:
    df = df.loc[~df["round_skipped"]]
    
# added:
df["won"] = df["won"].astype(int)

df

df.describe()

* Weak, but clear not extreme bias in favor of "first" cards, or in middle of screen?

df.corrwith(df["won"])

df.groupby(["ID_index"])["won"].mean()

#### normalize \____ 's 
* Rememebr to do this if evaluating
* May not be needed!
    * We also do this for consistency when doing find and replace , and for cases where there's no ___ in the text

df["has_"] = df["black_card_text"].str.contains("_{2,}")
### filter out cases with 2 blanks to fill ? 

df['black_card_text'] = df['black_card_text'].str.replace("_{2,}","__")
## https://stackoverflow.com/questions/47696401/replace-character-of-column-value-with-string-from-another-column-in-pandas


### Pick 2s
* Keep only subset for free text, and self join is slow.... 
* Keep only sample of combinations, as otherwise we get huuuge # combinations

# df["black_card_pick_num"].describe() ## max 2, min 1

df_doubles = df.loc[df["black_card_pick_num"]>1].filter(['black_card_text', 'white_card_text',"won","fake_round_id"],axis=1).drop_duplicates().copy()

df = df.loc[df["black_card_pick_num"]<2] ## keep non doubles only

# df[["id_black","id_white"]] = ord_enc.fit_transform(df[["black_card_text","white_card_text"]])
df["id_white"] = ord_enc.fit_transform(df["white_card_text"].values.reshape(-1, 1))

if not DROP_DOUBLES:
    df_doubles["black_card_text"] = df_doubles["black_card_text"].str.replace("(PICK 2)","",case=False,regex=False)

    df_doubles["black_card_text"] = df_doubles["black_card_text"].str.replace("__.","__",case=False,regex=False)
    df_doubles.sort_values(['black_card_text',"won"],inplace=True,ascending=False)
    # display(df_doubles)
    print(df_doubles.nunique())
    print(df_doubles.drop_duplicates(['black_card_text', 'white_card_text']).shape[0],"unique double combinations")
    print("Rows:",df_doubles.shape[0])

    ### biased sample - winning combinations + some others (otherwise, we have too many possibles..)
    df_doubles_1 = df_doubles.loc[df_doubles["won"]==True]
    df_doubles_2 = df_doubles.loc[df_doubles["won"]==False].groupby(["black_card_text","fake_round_id"]).sample(2)
    df_doubles = pd.concat([df_doubles_1,df_doubles_2]).drop_duplicates(['black_card_text', 'white_card_text'])
    print("after biased sample: Rows:",df_doubles.shape[0])
    df_doubles.set_index(["black_card_text","fake_round_id"],inplace=True)

    # ## keep first, last occ per all the white cards - biased sample, 
    # df_doubles = pd.concat([df_doubles.drop_duplicates(keep="first"),
    #                         df_doubles.drop_duplicates(subset=["white_card_text"], keep="last")
    #                        ]).drop_duplicates()
    # df_doubles = df_doubles.drop_duplicates(subset=["white_card_text"],keep="first")

    ### could filter for mirror images ? 
    ### https://stackoverflow.com/questions/24676705/pandas-drop-duplicates-if-reverse-is-present-between-two-columns
    # df['check_string'] = df.apply(lambda row: ''.join(sorted([row['InteractorA'], row['InteractorB']])), axis=1)

    # df_doubles.drop(columns="won",errors="ignore",inplace=True)
    print("Rows (after dropping):",df_doubles.shape[0])
    df_doubles

%%time
if not DROP_DOUBLES:
    ### slow merge + check string
    # df_doubles = df_doubles.merge(df_doubles,on=["black_card_text"])
    ## self join
    df_doubles = df_doubles.join(df_doubles["white_card_text"],lsuffix="_x",rsuffix="_y")
    df_doubles = df_doubles.loc[df_doubles["white_card_text_x"]!=df_doubles["white_card_text_y"]]
    # ## drop mirror images (maybe relevant, but too redundant)
    df_doubles['check_string'] = df_doubles.apply(lambda row: ''.join(sorted([row['white_card_text_x'], row['white_card_text_y']])), axis=1)

    df_doubles = df_doubles.reset_index().drop_duplicates(subset=["check_string"])
    df_doubles =df_doubles.sort_values(['black_card_text',"won"],ascending=False).drop(columns=["won","check_string","fake_round_id"],errors="ignore").drop_duplicates()
    print(df_doubles.shape) ## 200 million by default


    %%time
    ## remove . at end of text
    df_doubles["white_card_text_x"] = df_doubles["white_card_text_x"].str.replace("\.$", '',regex=True)
    df_doubles['text'] = df_doubles.apply(lambda x:x['black_card_text'].replace("__", x['white_card_text_x'],1), axis=1)
    df_doubles['text'] = df_doubles.apply(lambda x:x['text'].replace("__", x['white_card_text_y'],1), axis=1)
    df_doubles['text'] = df_doubles['text'].str.replace("..",".",regex=False) ## fix double dots. Still levaes extra punct.. 
    # df_doubles = df_doubles[["black_card_text",'text']].drop_duplicates()
    print(df_doubles.shape)
    df_doubles

## df with all texts from data
white, black and combinations

df_text = pd.concat([df["black_card_text"],
                       df["white_card_text"],
#                     df_doubles['text'],df_doubles['black_card_text']
                    ]).drop_duplicates()

Fill in "punchlines" into blanks (___) (or add at end of text for cards without ___)

* NOTE: Previously I had tried a model with black and white text as 2 prompts (which meant the model needed to learn to fill in blanks); and a model of just the filled in answer (1 col). 
* Unclear what is "easier" to learn/model

* Note this doesn't handle cases with double blanks (it will just input the same punchline twice)

df["has_"] = df["black_card_text"].str.contains("_{2,}",regex=True)

df['black_card_text'] = df['black_card_text'].str.replace("_{2,}","__",regex=True)
## https://stackoverflow.com/questions/47696401/replace-character-of-column-value-with-string-from-another-column-in-pandas
## https://datascience.stackexchange.com/questions/39345/how-to-replace-a-part-string-value-of-a-column-using-another-column
# Remove characters from one column based on string of another column

## remove "." from end of text
df["white_card_text"] = df["white_card_text"].str.replace("\.$", '',regex=True)
df['text'] = df.apply(lambda x:x['black_card_text'].replace("__", x['white_card_text']), axis=1)

## add answer at end of text, in cases where no ___ in black card
df.loc[(df["has_"]==False),"text"] = df["text"] +" " + df["white_card_text"] #+ "."

# removed: 
# df['text'] = df['text'].str.replace("..",".",regex=False) ## fix double dots. Still levaes extra punct.. 

display(df)

* Add these to df_text as well

Save df_text to disk

df_text = pd.concat([df_text,df['text']]).drop_duplicates()
df_text = df_text.sample(frac=1)
df_text

# df_text.to_csv("df_text.csv.gz",index=False,compression="gzip")

##### Fit prior odds / encoder
* Use targetEncoder or WOE
* Do IPW either here, or later on data subset? 

assert df["won"].isna().sum()==0

df["won"].mean()

# ## for now, fit on just white cards, instead of black, white  (For black, we could just use mean_skipped as prior and multiply)
# ## seemed to work directly on text, origianlly? 
# df["prior_white"] = cat_encoder.fit_transform(X= df["white_card_text"].values# df["id_white"].values# pd.factorize(df.white_card_text)[0],#df["white_card_text"],
#                                               ,y=df["won"].astype(int).values,groups=df["fake_round_id"].values)

# # df["prior_white"] = cat_encoder.transform(df["white_card_text"]) ## catboost encoder - doesn't have min 0 (for those with 0 cases)
# ## rates for the whites (per unique white card)
# df.drop_duplicates(["white_card_text"])["prior_white"].describe(percentiles=[]).round(3)

# ## why nan values ??

### check white prior as baseline model

* 21% win rate by picking most popular answer, on FILTERED Data
* On _full_ 3 million games (including different outcomes for same pair, i.e no filtering by # picks), _it does not better than random! _

df.groupby(["ID_index"])["won"].mean()

### get min count/cooccurences of sentence pairs
* Filter train +_ TEST for min 2 occurrences
* Filter games/rounds where positive was removed

df["pair_count"] = df.groupby("text")["won"].transform("count") ## can be used to filter sentences occurring less than k times
df["sum_won"] = df.groupby("text")["won"].transform("sum")

print(df.shape[0])
df = df.loc[df["pair_count"]>=MIN_PAIR_FREQ]
print(df.shape[0])

### remove groups where positive was filtered out
( If `MIN_PAIR_FREQ` is 1, then this does nothing, all cards/games/jokes will be kept)

if MIN_PAIR_FREQ>1:
    print("Rounds before:",df["fake_round_id"].nunique())
    any_pos = df.groupby("fake_round_id")["won"].transform("max")>0
    df = df.loc[any_pos]
    print("Rounds After (filtering those with no pick):",df["fake_round_id"].nunique())
    
    ### filter round with too few cards left (we can have without all 10, but ensure no "1" only cases.... )
     ## clear drop between 8,9 (17K),10 (6k) games
    round_cards_count_mask = df.groupby("fake_round_id")["won"].transform("count")>=8
    df = df[round_cards_count_mask==True]
    print(df.shape[0],"rows after round level filter")
    print(df["fake_round_id"].nunique(), "rounds")

print(df.nunique())

df[df["sum_won"]>0]["sum_won"].describe().round(2)

df.nunique()

print("Possible jokes:",(df.black_card_text.nunique()*df.white_card_text.nunique()))
100*df.text.nunique() / (df.black_card_text.nunique()*df.white_card_text.nunique())

## Split train/test by groups
* Later-  filter train data by label (but not the test)
* split by round_id / games

df = df.filter(['fake_round_id', 'black_card_text', 'white_card_text',
                'won', 'text', "sum_won"],axis=1)
df = df.sort_values(["fake_round_id"],ascending=False) # , 'won'
df["won"] = df["won"].astype(int)
print("mean won:",df["won"].mean())
print(df.shape[0])

## Alt: split by unique cards
* unique / novel white, black card combinations
* Randomly select

all_blacks = df["black_card_text"].unique()
# print(len(all_blacks))
# test_blacks = np.random.choice(all_blacks,size=int(len(all_blacks)*0.05), replace=False)

all_whites = df["white_card_text"].unique()

test_whites = np.random.choice(all_whites,size=int(len(all_whites)*0.2), replace=False)
print("all_whites",len(all_whites),"test whites:",len(test_whites))

card_mask = (df["white_card_text"].isin(test_whites))
df_test_cards = df.loc[card_mask].copy()
df_train_cards = df.loc[~card_mask].copy()

print("train rows",df_train_cards.shape[0])
print("test rows",df_test_cards.shape[0])
print("df_train_cards mean won",df_train_cards.won.mean())
print("df_test_cards mean won",df_test_cards.won.mean())

if SAVE_TRAIN_DATA:
    df_train_cards.to_parquet("cah_train_cardsplit_games.parquet")
    df_test_cards.to_parquet("cah_test_cardsplit_games.parquet")

### "Normal" round-level train/test split:

* ALT - below - take last k rounds

df["fake_round_id"].nunique()

df = df.sample(frac=1).sort_values("fake_round_id")

df.shape[0] # 2,446,360 rows  , and 298,955 games

test_size = df.shape[0]//5 ## 20% 
## round to nearest 10
test_size = int(math.ceil(test_size / 10.0)) * 10

df_test = df.tail(test_size)
print(df_test.shape[0])
display(df_test.head(12))

#### subset df to train

df_train = df.iloc[:-test_size]
df_train.shape[0]

## # Initialize the GroupShuffleSplit.
# gss = GroupShuffleSplit(n_splits=1, test_size=0.2,random_state=42)

# # Get the indexers for the split.
# idx1, idx2 = next(gss.split(df, groups=df["fake_round_id"].values))

# # Get the split DataFrames.
# df, df_test = df.iloc[idx1], df.iloc[idx2]

# print("df (train)",df.shape[0])
# print(df["fake_round_id"].nunique())
# print("test")
# print("df_test",df_test.shape[0])
# print(df_test["fake_round_id"].nunique())

# df_train = df

### save to disk

if SAVE_TRAIN_DATA:
#     df.to_parquet("cah_train_min3_games.parquet")
#     df_test.to_parquet("cah_test_min3_games.parquet")
#     df.to_parquet("cah_train_games.parquet")
    
    df_train.to_parquet("cah_train_games.parquet")
    df_test.to_parquet("cah_test_games.parquet")

#### Whites prior: 
* 20.5% accuracy on test set!!;
* 23% after min 3 filtering

# df_white_prior = df.groupby(["white_card_text"], as_index=False)["won"].mean().rename(columns={"won":"white_prior"}).set_index("white_card_text")
df_white_prior = df_train.groupby(["white_card_text"], as_index=False)["won"].mean().rename(columns={"won":"white_prior"}).set_index("white_card_text")
df_test = df_test.join(df_white_prior,on="white_card_text",how="left")
prior = df_test["white_prior"].mean()
# df_test["white_prior"] = df_test["white_prior"].fillna(prior)
df_test["white_prior"] = df_test["white_prior"].fillna(0.1)
print(df_test.shape[0],"test shape after join")
print("Prior Acc @1:",df_test.sort_values("white_prior",ascending=False).groupby("fake_round_id").head(1)["won"].mean())
## alkt: 
print("alt dedup Prior Acc @1:",df_test.sort_values("white_prior",ascending=False).drop_duplicates(subset=["fake_round_id"])["won"].mean())
# print("Prior Acc @1:",df_test.sort_values("white_prior",ascending=False).groupby("fake_round_id").head(1).groupby("fake_round_id")["won"].max().mean()) ## res identical to above
print("Prior Acc @2:",df_test.sort_values("white_prior",ascending=False).groupby("fake_round_id").head(2).groupby("fake_round_id")["won"].max().mean())
print("Prior Acc @3:",df_test.sort_values("white_prior",ascending=False).groupby("fake_round_id").head(3).groupby("fake_round_id")["won"].max().mean())

### keep 1s (for train) ?
* MAY want to filter - e.g. min # cases of it being picked (+- keeping that for use for sample weight)
* Keep all data for now, (to allow for grouping by rounds if we want, since this is a ranking problem) 

* temp

df_tr = df[['fake_round_id',"black_card_text","white_card_text","won"]].sort_values("won",ascending=False).copy()
df_tr =df_tr.drop_duplicates(subset=['fake_round_id',"black_card_text","white_card_text"])
df_tr["white_counts"] = df_tr.groupby("white_card_text")["won"].transform("size")
print(df_tr.nunique())
df_tr

df_tr["white_counts"].describe().round(1)

## Sentence transformers model(s)
* Sentence bert (NLI? semtantic similarity?) 
https://www.sbert.net/docs/training/overview.html
https://www.sbert.net/docs/quickstart.html  , https://github.com/UKPLab/sentence-transformers
* Load data from disk ? 

* list of pretrained sentencebert models: (All are >300 M! slow download) https://docs.google.com/spreadsheets/d/14QplCdTCDwEmTqrn1LH4yrbKvdogK4oQvYO1K1aPR5M/edit#gid=0

* `stsb-distilroberta-base-v2` - 305M sized model - semantic similarity
* `nli-distilroberta-base-v2` - NLI 
* `average_word_embeddings_glove.6B.300d`, `average_word_embeddings_glove.840B.300d` - glove/w2v embeddings

from sentence_transformers import SentenceTransformer, SentencesDataset, InputExample, losses, evaluation
import torch
torch.cuda.is_available()
from torch.utils.data import DataLoader
import torch.utils.data as data_utils
## https://stackoverflow.com/questions/50307707/convert-pandas-dataframe-to-pytorch-tensor
from transformers import AutoTokenizer, AutoModel
## example from : https://www.sbert.net/docs/training/overview.html
from sentence_transformers import SentenceTransformer, SentencesDataset, InputExample, losses 


## pretrained weights - download vis program often fails, easier to download externally from ftp
# model = SentenceTransformer('stsb-distilroberta-base-v2') ## was distilbert-base-nli-mean-tokens

# model = SentenceTransformer("nli-distilroberta-base-v2",device="cuda")
# model = SentenceTransformer("paraphrase-MiniLM-L12-v2") ## "paraphrase-MiniLM-L12-v2"
model = SentenceTransformer("all-MiniLM-L12-v2"
# model = SentenceTransformer("./stsb-distilroberta-base-v2",device="cuda")

# model = SentenceTransformer("./nli-mpnet-base-v2",device="cuda")

model.max_seq_length

# len(model.tokenize(["ad and "])["input_ids"])

model.get_sentence_embedding_dimension() # 768

#### redo above, but with real data


num_epochs = 2
TEST_SIZE = df.shape[0]//8 #4500
print(TEST_SIZE)

df.shape

df.drop_duplicates(['black_card_text', 'white_card_text']).shape[0]

## is list correct? error otherwise? 
train_examples = list(df.iloc[TEST_SIZE:].apply(lambda row: InputExample(texts=[row["white_card_text"], row["black_card_text"]],
                                                                         label=row["picks"]),axis=1)) # picks
test_examples =  list(df.iloc[0:TEST_SIZE].apply(lambda row: InputExample(texts=[row["white_card_text"], row["black_card_text"]], 
                                                                          label=row["picks"]),axis=1)) # picks


# dev_eval = evaluation.BinaryClassificationEvaluator(sentences1=df.iloc[0:2500,0],
#     sentences2= df.iloc[0:2500,1],
#     labels=df.iloc[0:2500,2],show_progress_bar=True,batch_size=256)
dev_eval = evaluation.BinaryClassificationEvaluator(sentences1=list(df.iloc[0:TEST_SIZE,0].values),
    sentences2= list(df.iloc[0:TEST_SIZE,1].values),
    labels=list(df.iloc[0:TEST_SIZE,2].values),show_progress_bar=True,batch_size=256)

## https://github.com/UKPLab/sentence-transformers/blob/master/examples/training/nli/training_nli.py 

## evaluator example 
train_dataloader = DataLoader(train_examples, shuffle=True, batch_size=256)


%%time
# train_loss = losses.CosineSimilarityLoss(model) # ORIG

# train_loss = losses.SoftmaxLoss(model,
#     sentence_embedding_dimension= model.get_sentence_embedding_dimension(),
#     num_labels= 2,
#     concatenation_sent_rep= True,
#     concatenation_sent_difference = True,
#     concatenation_sent_multiplication = True) # ALT - for classification https://www.sbert.net/docs/package_reference/losses.html#softmaxloss

train_loss = losses.CosineSimilarityLoss(model)
#Tune the model
model.fit(train_objectives=[(train_dataloader, train_loss)], 
          epochs=num_epochs, 
#           warmup_steps=100, # 100
          warmup_steps = math.ceil(len(train_dataloader) * 0.05*3), #5% of train data for warm-up
          evaluator=dev_eval,
          use_amp=True,
          optimizer_params= {'lr': 3e-04}, # 2e-05
          output_path="./output",
          save_best_model = False,
         )

model.evaluate(dev_eval)

#### Evaluate model on original data, game by game.
* Note: should use proper test set , disjoint

# df = pd.read_parquet('cah_lab_data.parquet')
df_test = df_sample.loc[df_sample["black_card_pick_num"]==1].sort_values("fake_round_id")[["fake_round_id","black_card_text","white_card_text","won"]].copy()

df_test = df_test.head(4800)
df_test

* if using just black/white, could do this for unique cards - ~1000 x less compute...

%%time
## https://www.sbert.net/docs/usage/semantic_textual_similarity.html

sentences1 = df_test["black_card_text"].values
sentences2 = df_test["white_card_text"].values
embeddings1 = model.encode(sentences1,batch_size=256,normalize_embeddings=False, convert_to_tensor=True)

embeddings2 = model.encode(sentences2,batch_size=256,normalize_embeddings=False, convert_to_tensor=True)

scores = []
#Compute cosine-similarits
cosine_scores = util.pytorch_cos_sim(embeddings1, embeddings2)

#Output the pairs with their score
for i in range(len(sentences1)):
#     print("{} \t\t {} \t\t Score: {:.4f}".format(sentences1[i], sentences2[i], cosine_scores[i][i]))
    scores.append(float(cosine_scores[i][i]))
df_test["cos_scores"] = scores
df_test.sort_values(["fake_round_id","cos_scores"],inplace=True,ascending=True)
print("top1 accuracy:",df_test.groupby("fake_round_id").head(1)["won"].mean())

df_test.sort_values(["fake_round_id","cos_scores"],inplace=True,ascending=False)
print("top1 accuracy, reverse sort:",df_test.groupby("fake_round_id").head(1)["won"].mean())
df_test
